refactor(retention): make core/retention a package rather than a second root leaf
`tests/quality/test_no_flat_core_regrowth.py` went red at 104 leaves against a
pin of 103: `retention_task_results.py` landed at the root of `src/teatree/core/`
beside the `retention.py` that already owns that concern. Two root modules for
one concept is the #3854 shape, so the pin is NOT raised — retention becomes the
subpackage instead, and the pin is LOWERED to 102 because two leaves collapse
into one package entry.

Chose the package over folding the two files into one. Folding would put the
`django_tasks_db` import inside the lane orchestrator, which is the seam the
split exists to hold: `retention/prune.py` owns the lanes and never imports a
third-party table's library, `retention/task_results.py` is the adapter onto
`django_tasks_db`'s own shipped `prune_db_task_results`. That is also the shape
the surrounding code already uses — `core/cleanup/`, `core/evidence/`,
`core/intake/` and `core/review/` are all docstring-only `__init__` packages
whose members are imported by submodule path.

Tests move with the modules to keep the mirror exact
(`tests/teatree_core/retention/test_prune.py` + `test_task_results.py`).
